Add time slider that controls the program.
This required a rather massive refactoring of the way simulation works.
Now, whenever the program is changed, it is internally fully simulated
and all game states are stored in a big ol array. "Executing" the program
merely displays the various items in this array, and the time slider
seeks to particular positions in the array directly.