Add functions and queries to the ast.
Add support to the ast, parser and simulator for function and queries.
Queries are to expressions as commands are to statements,
that is, they are questions that can be asked of the robot in an expression.
Robot has hacked in support for queries, though it's clear that there
needs to be a bit o' refactoring to properly support them.